Use meaningful name for mk_property_methods
The original ParsedFile provider method had a `mk_resource_methods`
method that generated attr_accessors for parameters and properties. To
prevent parameters from being treated as properties for the sake of
sanity this method will only generate accessors for actual properties.
However, this behavior changes the semantics of the method, so it was
renamed to reflect this behavior change.