(filemapper-#3) Add :unlink_empty_files
If a file has no provider instances associated with it, it frequently
makes sense to remove the file entirely. This adds an
`unlink_empty_files` accessor to Filemapper. If it's set to true and a
provider returns the empty string as the contents of a file, then the
file will be backed up and removed.